Bamboo Plan
Creation of a specific command to validate whether the execution has finished.
Must be used in the Final Part of the Bamboo CI
